requirements.txt
datasets
huggingface_hub<1.0.0
requests
beautifulsoup4
google-genai
pubchempy
rdkit
gradio
